M03 Programming Assignment 1
This assignment rewrites the Circle class to extend GeometricObject and implement the Comparable interface. The equals method in the Object class gets overridden. Two Circle objects are equal if their radii are the same.
The UML diagram that involves Circle, GeometricObject, and Comparable are in a word file, which contains the URL of this assignment.