M03 Programming Assignment 2
In this assignment, I redesign and implement the Rational class in Listing 13.13 using BigInteger for the numerator and denominator. This is a test program that prompts the user to enter two rational ­numbers and displays what happens when they two get added, subtracted, multiplied and divided.
It also displays the big decimal for the second integer.